# api/openapi/specs.yaml
openapi: 3.0.3
info:
  title: Feed Backend API
  version: 1.0.0
  description: |
    API para sistema de feed tipo Instagram con arquitectura Publisher/Subscriber.
    Soporta autenticación JWT, paginación basada en cursor, operaciones idempotentes
    y entrega de notificaciones en tiempo real.
  contact:
    name: Backend Team
    email: backend@feed.local
servers:
  - url: http://localhost:8080
    description: Gateway de Desarrollo Local
  - url: https://api.feed.local
    description: Entorno de Producción
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado exitosamente
          headers:
            Idempotent:
              schema: { type: boolean }
              description: true si la respuesta fue servida desde caché idempotente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Autenticar usuario
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens emitidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
...
  /feed:
    get:
      tags: [Feed]
      summary: Obtener timeline principal
      operationId: getFeed
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de posts paginada
          headers:
            X-Next-Cursor:
              schema: { type: string }
              description: Cursor para la siguiente página
            Cache-Control:
              schema: { type: string, example: "max-age=30, stale-while-revalidate=60" }
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeedEntry'
                  next_cursor:
                    type: string
                    nullable: true
                    description: null si no hay más páginas

  # ─────────────────────────────────────────────────────────────
  # INTERACTIONS
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT corto (15m) obtenido via /auth/login
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string, format: base64 }
      description: Cursor opaco retornado en X-Next-Cursor
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 30, pattern: '^[a-zA-Z0-9_\.]+$' }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, format: password }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthResponse:
      type: object
      required: [user_id, access_token, expires_in]
      properties:
        user_id: { type: string, format: uuid }
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, format: int64, description: "Segundos hasta expiración" }
...
    FeedEntry:
      type: object
      properties:
        post_id: { type: string, format: uuid }
        author: { type: object, properties: { id: { type: string }, username: { type: string } } }
        preview: { type: object, properties: { caption: { type: string, nullable: true }, media_count: { type: integer } } }
        metrics: { type: object, properties: { likes: { type: integer }, comments: { type: integer } } }
        created_at: { type: string, format: date-time }
        cursor: { type: string, description: "Cursor estable para este item" }
...
    Error:
      type: object
      properties:
        type: { type: string, format: uri, example: "https://api.feed.local/errors/validation" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-path }
        request_id: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    ValidationError:
      description: Error de validación de payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnauthorizedError:
      description: Token inválido, expirado o ausente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ConflictError:
      description: Conflicto de unicidad o estado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - BearerAuth: []